pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "A production-ready FastAPI backend for Next.js"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.27",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.26.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-logging>=0.44b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3 